feat: 헤더에 있던 스크롤이벤트 APP으로 옮겨서 전체 페이지 공통으로 사용
