+++ src/components/common/Header.jsx
@@ -8,9 +8,8 @@ import * as C from '@/style/CommonContents';
 import { LogoBlack, LogoWhite } from '@/assets/img/logo';
 import { TransEn, TransKo } from '@/assets/img/icons';
 
-const Header = ({ bgColor }) => {
-    const [scrollHeader, setScrollHeader] = useState(false); //헤더 배경 검정색인지 아닌지
-    const [logoColor, setLogoColor] = useState(LogoWhite);
+const Header = ({ bgColor, pageName, scrollHeader }) => {
+    const [logoColor, setLogoColor] = useState(LogoWhite); //로고 컬러 상태
 
     //배경 색상에 따른 로고 색상 변경 스위치
     const switchBg = () => {
@@ -26,27 +25,13 @@ const Header = ({ bgColor }) => {
         }
     };
 
-    //스크롤시 배경 on
-    const handleScroll = () => {
-        const scrollPosition = window.scrollY; //스크롤 위치
-        scrollPosition > 50 ? setScrollHeader(true) : setScrollHeader(false);
-    };
     //배경 색상 변경 감지
     useEffect(() => {
         switchBg();
     }, [bgColor]);
 
-    //스크롤 이벤트 감지
-    useEffect(() => {
-        handleScroll();
-
-        window.addEventListener('scroll', handleScroll);
-        return () => {
-            window.removeEventListener('scroll', handleScroll);
-        };
-    }, []);
     return (
-        <C.HeaderCont scrollHeader={scrollHeader} setScrollHeader={setScrollHeader} bgColor={bgColor}>
+        <C.HeaderCont scrollHeader={scrollHeader} pageName={pageName} bgColor={bgColor}>
             <C.HeaderInner>
                 <C.HeaderLogo>
                     <Link to={'/'}>
